test(mla): fold sharding regression into existing suite
