+++ test/srt/run_suite.py
@@ -344,11 +344,6 @@ suites = {
             0.2,
             runner="pytest",
         ),
-        TestFile(
-            "python/sgl_jax/test/test_mla_backend_sharding.py",
-            0.2,
-            runner="pytest",
-        ),
         TestFile("python/sgl_jax/test/mem_cache/test_req_to_token_pool.py", 1),
         TestFile("python/sgl_jax/test/mem_cache/test_hybrid_req_to_token_pool.py", 1),
         TestFile("python/sgl_jax/test/mem_cache/test_swa_allocator.py", 1),
